Move sanitized heap to 'extensions'.
